# tuma-platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuma API",
    "description": "Backend API for the Tuma commerce + delivery platform",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Platform sign-in (merchant operators and admins) with email + password.\nSets the two session cookies (15-min access JWT + 30-day opaque refresh);\nthe body is empty. Authorization is resolved from profiles — an account\nwith neither an admin profile nor a merchant membership cannot enter the\nplatform, and the answer never reveals which part failed.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Session cookies set (tuma-auth-token + tuma-auth-refresh)"
          },
          "401": {
            "description": "One generic error for every failure path"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "otp_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Always the same response — reveals nothing about the phone number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid phone number"
          }
        }
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "otp_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Code valid — account and customer profile fetched or created, token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change the caller's own password. Merchant operators and admins only in\npractice — customers have no password (OTP accounts), which is a 400.",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Account has no password (OTP sign-in)"
          },
          "401": {
            "description": "Not authenticated, or current password is wrong"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "The authenticated account with its profiles and memberships",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        }
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Updates the caller's own profile name — the customer profile when there\nis one, else the admin profile. Merchant-operator accounts have no\neditable name in V1 (their identity on the wing is the business).",
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty name, or no editable profile"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordInput": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
...
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "description": "The authenticated account with its authorization context. Deliberately\nnot the raw domain rows — responses never carry `password_hash`. The\naccount itself is roleless; what it can do is visible in the profiles\nand memberships below, all resolved fresh per request.",
        "required": [
          "id",
          "created_at",
          "merchant_memberships"
        ],
        "properties": {
          "admin": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProfileResponse"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "customer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProfileResponse"
              }
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "merchant_memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MembershipResponse"
            }
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "MembershipResponse": {
        "type": "object",
        "required": [
          "merchant_id",
          "merchant_name",
          "role",
          "merchant_status"
        ],
        "properties": {
          "merchant_id": {
            "type": "string",
            "format": "uuid"
          },
          "merchant_name": {
            "type": "string"
          },
          "merchant_status": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "store_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "`null` = the whole business; a value = this member is scoped to that\none store."
          }
        }
      },
...
      "OtpRequestInput": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string"
          }
        }
      },
      "OtpRequestResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "OtpVerifyInput": {
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "OtpVerifyResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MeResponse"
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateMeInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}